refactor(core, dsp, docs): polish architecture docs, add Biquad & Mailbox hardware examples, fix static analysis & Doxygen coverage
- Clarify Cooperative-by-Design philosophy in README.md and docs (native state machines, PT_DEFER, zero heap allocation)
- Differentiate 2-byte raw protothread continuation (syn_pt) vs ~16-28 byte scheduled task descriptors (syn_task)
- Add BiquadFilter (Q16.16 2nd-order Butterworth lowpass) and TaskMailbox (SPSC lock-free IPC) hardware examples
- Fix protothread Duff's device continuation variable lifetime in FftSpectrumAnalyzer
- Fix strict C99 static analysis issues in syn_websocket.c, syn_sha256.c, syn_fat.c, and syn_wg.c
- Add syn_crypto and syn_util Doxygen defgroups to syntropic.h for 100% zero-warning Doxygen API coverage
- Verify 100/100 Unity unit test modules PASS across 8-bit AVR and 32-bit ARM targets

+++ src/syntropic/util/syn_sha256.c
@@ -182,12 +182,12 @@ void syn_sha256_update(SYN_SHA256 *ctx, const void *data, size_t len)
     const uint8_t *p = (const uint8_t *)data;
 
     /* Track total length in bits (split into hi/lo for >4GB messages) */
-    uint32_t len_bits_lo = (uint32_t)(len << 3);
+    uint32_t len_bits_lo = (uint32_t)((uint64_t)len << 3);
     ctx->total_len_lo += len_bits_lo;
     if (ctx->total_len_lo < len_bits_lo) {
         ctx->total_len_hi++;  /* carry */
     }
-    ctx->total_len_hi += (uint32_t)(len >> 29);
+    ctx->total_len_hi += (uint32_t)((uint64_t)len >> 29);
 
     /* Fill partial buffer first */
     if (ctx->buf_len > 0) {
